Fix session improvements
